Shrink pathfinding memory footprint by ~770 KB per workspace
Two contained changes that drop per-query and per-volume memory without
changing search semantics:

1. Fold AStarWorkspace.closed (Uint8Array, 384 KB) into the existing
   gen Int32Array. Each query now owns a pair of consecutive generation
   tickets — `gen` for relaxed/open cells, `gen + 1` for closed. genTick
   advances by 2 per reset() so the two markers can never collide with
   another query's. Saves 384 KB per workspace.

2. Shrink VolumeGrid.topY from Int16Array to Uint8Array. Voxel y is in
   0..WORLD_Y-1 = 0..191, so 255 fits as the all-air sentinel (replacing
   -1). Saves another 384 KB on the shared volume grid; AStar's
   step-climb gate updates from `>= 0` to `!== 255`.

Typecheck clean, all 158 tests pass, vite build succeeds.

+++ src/path/VolumeNav.ts
@@ -31,12 +31,12 @@ export interface VolumeNavBuffers {
   bedrock: Uint8Array;
   digCost: Uint8Array;
   /**
-   * Voxel-y of the highest solid voxel inside each cell, or -1 if all-air.
+   * Voxel-y of the highest solid voxel inside each cell, or 255 if all-air.
    * Replaces the old `surfaceConnected` flood-fill bitmap (the new pathfinder
    * uses per-unit grids instead, so the sealed-cave gate is no longer needed
    * — units route into caves only when their body actually fits the corridor).
    */
-  topY: Int16Array;
+  topY: Uint8Array;
   /**
    * Same shape as the old buffer for callers that still write to it. Always 0
    * after build; kept so any legacy code that consults it sees a defined
